+++ tests/test_account.py
@@ -1,0 +1,94 @@
+import responses
+from util import *
+
+import nexmo
+import platform
+
+@responses.activate
+def test_get_balance(client, dummy_data):
+    stub(responses.GET, 'https://rest.nexmo.com/account/get-balance')
+
+    assert isinstance(client.get_balance(), dict)
+    assert request_user_agent() == dummy_data.user_agent
+
+
+@responses.activate
+def test_application_info_options(dummy_data):
+    app_name, app_version = 'ExampleApp', 'X.Y.Z'
+
+    stub(responses.GET, 'https://rest.nexmo.com/account/get-balance')
+
+    client = nexmo.Client(key=dummy_data.api_key, secret=dummy_data.api_secret, app_name=app_name,
+                          app_version=app_version)
+    user_agent = '/'.join(
+        ['nexmo-python', nexmo.__version__, platform.python_version(), app_name, app_version])
+
+    assert isinstance(client.get_balance(), dict)
+    assert request_user_agent() == user_agent
+
+
+@responses.activate
+def test_get_country_pricing(client, dummy_data):
+    stub(responses.GET, 'https://rest.nexmo.com/account/get-pricing/outbound')
+
+    assert isinstance(client.get_country_pricing('GB'), dict)
+    assert request_user_agent() == dummy_data.user_agent
+    assert 'country=GB' in request_query()
+
+
+@responses.activate
+def test_get_prefix_pricing(client, dummy_data):
+    stub(responses.GET, 'https://rest.nexmo.com/account/get-prefix-pricing/outbound')
+
+    assert isinstance(client.get_prefix_pricing(44), dict)
+    assert request_user_agent() == dummy_data.user_agent
+    assert 'prefix=44' in request_query()
+
+
+@responses.activate
+def test_get_sms_pricing(client, dummy_data):
+    stub(responses.GET, 'https://rest.nexmo.com/account/get-phone-pricing/outbound/sms')
+
+    assert isinstance(client.get_sms_pricing('447525856424'), dict)
+    assert request_user_agent() == dummy_data.user_agent
+    assert 'phone=447525856424' in request_query()
+
+
+@responses.activate
+def test_get_voice_pricing(client, dummy_data):
+    stub(responses.GET, 'https://rest.nexmo.com/account/get-phone-pricing/outbound/voice')
+
+    assert isinstance(client.get_voice_pricing('447525856424'), dict)
+    assert request_user_agent() == dummy_data.user_agent
+    assert 'phone=447525856424' in request_query()
+
+
+@responses.activate
+def test_update_settings(client, dummy_data):
+    stub(responses.POST, 'https://rest.nexmo.com/account/settings')
+
+    params = {'moCallBackUrl': 'http://example.com/callback'}
+
+    assert isinstance(client.update_settings(params), dict)
+    assert request_user_agent() == dummy_data.user_agent
+    assert 'moCallBackUrl=http%3A%2F%2Fexample.com%2Fcallback' in request_body()
+
+
+@responses.activate
+def test_topup(client, dummy_data):
+    stub(responses.POST, 'https://rest.nexmo.com/account/top-up')
+
+    params = {'trx': '00X123456Y7890123Z'}
+
+    assert isinstance(client.topup(params), dict)
+    assert request_user_agent() == dummy_data.user_agent
+    assert 'trx=00X123456Y7890123Z' in request_body()
+
+
+@responses.activate
+def test_get_account_numbers(client, dummy_data):
+    stub(responses.GET, 'https://rest.nexmo.com/account/numbers')
+
+    assert isinstance(client.get_account_numbers(size=25), dict)
+    assert request_user_agent() == dummy_data.user_agent
+    assert 'size=25' in request_query()
